requirements.txt
flask==3.0.0
watchdog==4.0.0
pandas==2.1.4
pillow==10.1.0
flask-cors==4.0.0
